Generate a random sample of points on the unit sphere

Random sample

Random sample with minimum exlusion zone ("hard-core process")

Fibonacci coverage of a sphere

sample_random(N)

sample_hc(N, exclusion = 0.1, maxiter = 200L, k = 30L)

sample_fibonacci(N = 301)

Arguments

N

number of points

exclusion

minimum distance allowed between points

maxiter

maximum number of iterations

k

number of extra new points to try at each iteration

Value

3xN matrix

3xN matrix

Details

Produces a set of points that covers rather uniformly the unit sphere with N points with a spiral-like pattern based on a Fibonacci sequence

Functions

sample_random: random sample

sample_hc: random sample with exclusion zone

sample_fibonacci:

Examples

sample_random(10)

#>            [,1]       [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,] -0.5789046 -0.3840421  0.2932116 -0.5138034 -0.97967514  0.7758083
#> [2,] -0.3218179  0.6556350  0.2151799 -0.7790770 -0.04557996 -0.1662954
#> [3,]  0.7492013 -0.6501187 -0.9315173 -0.3592285 -0.19534352 -0.6086603
#>             [,7]       [,8]       [,9]       [,10]
#> [1,]  0.98111807  0.1207429  0.1046551 -0.04356523
#> [2,] -0.01370228 -0.4731341  0.9692769  0.30580892
#> [3,] -0.19292377 -0.8726771 -0.2225974  0.95109567

sample_hc(10)

#>            [,1]       [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,]  0.1884051 0.09120063 -0.3145843 -0.9739656  0.04190463 -0.4219582
#> [2,] -0.7358565 0.98592964 -0.8905525  0.1196866  0.78587499 -0.1413544
#> [3,]  0.6503988 0.14008990 -0.3285618  0.1925256 -0.61696394  0.8955279
#>            [,7]       [,8]       [,9]      [,10]
#> [1,] 0.97848548 0.33253179  0.3376169 -0.9752212
#> [2,] 0.18801019 0.93886355  0.8306250 -0.1938582
#> [3,] 0.08496082 0.08920679 -0.4428057 -0.1065951