spectrum_oa

Orientation-averaged spectrum

spectrum_oa(cluster, material, medium = 1.33, quadrature = c("gl", "qmc",
  "random", "cheap"), Nq = 100, iterative = FALSE, precision = 0.001,
  Qmax = 10000, dN = Nq, method = c("solve", "cg", "oos"), Nsca = 50,
  maxiter = 30, tol = 1e-04, progress = FALSE, verbose = TRUE)

Arguments

cluster

cluster (list)

material

material

medium

refractive index medium

quadrature

quadrature method, using either Gauss Legendre quadrature (default), Quasi Monte Carlo, random, or "cheap" (3 Axes)

Nq

number of integration points

iterative

logical, increase N until convergence (QMC only)

precision

relative diff between two runs (QMC only)

Qmax

maximum N if convergence not attained (QMC only)

dN

iterative increase in N (QMC only)

method

linear system (solve), conjugate-gradient (cg), or order-of-scattering (oos)

Nsca

quadrature points for scattering cross-section

maxiter

integer termination of iterative solver

tol

double, tolerance of iterative solver

progress

print progress lines

verbose

display messages

Details

OA spectrum

References

Y. Okada, Efficient numerical orientation quadrature of light scattering properties with a quasi-Monte-Carlo method, Journal of Quantitative Spectroscopy and Radiative Transfer, Volume 109, Issue 9, June 2008, Pages 1719-1742.